+++ owoow.Core/RNG/Generators/Overworld/Symbol.cs
@@ -117,7 +117,7 @@ public class Symbol
                     DexRec = GenerateDexRecActivation(ref rng);
                     if (DexRec > 49 && config.IsDexRecActive)
                     {
-                        DexRecSlot = (short)config.DexRecSlots[GenerateEncounterSlot(ref rng, 4)];
+                        DexRecSlot = config.DexRecSlots[GenerateEncounterSlot(ref rng, 4)];
                         var DexRecMatchingSpecies = ActiveTable.Where(enc => enc.Value.DevId == DexRecSlot);
                         if (DexRecMatchingSpecies.Any())
                         {
